cleaned up packing list code and traveler form
